Simplified iam tests cases to just validate json
It would make sense to validate the exact policy, however, policies
do change significantly and the only real changes are the content.

We also cannot determine if the policy generated is in fact legal
or represents what policies we'd like to enforce. Since that's the
case, i think we just ensure that generated policy is simply
valid json.